Preserve name and doc string of decorated function.
When decorating a function it is replaced by the decorator wrapper function. Meaning that packages like Sphinx that inspect functions would only see the name and doc string of the decorator, instead of the name and doc string of the function itself. When using functools.wraps the meta data of the original function is preserved.