Do not try to synchronize the cache at deletion-time in Pickle backend.
The call of __del__ is not predictable. Using __del__ to invoke a synchronization results in an unpredictable state. It is not clear whether or not the cache was written successfully to the file, because the cPickle module could have already been deleted at that time.
  
See http://docs.python.org/reference/datamodel.html#object.__del__