Use module decorator, add configurable default arguments, configurable hash function.
When decorating a function it is substituted by the decorator function. Therefore when accessing the name or doc string of the decorated function the name and doc string of the decorator is returned instead. The module decorator allows to almost transparently decorate functions, preserving the name, doc string and argument specification used by modules like Sphinx. The current available module decorator is added to the code-base as a fall-back in case it is not installed on the users system.

The default values of all parameters of the main memoizer decorator are now configurable by defining them in the config module. This allows to globally change default values (e.g. the used backend) when using memoizer.

A new hash argument for the memoizer decorator allows to override the default function that computes the hash-key when looking up the cache.

Tests have been updated to reflect the changes.