Remove path='auto' in FunctionCache.
It is not necessary anymore to explicitly set path to 'auto'. Now, a path that evaluates to False is invoking the replacement with a default path. This only works because 'backend' can be (and is by default) 'None' since the last revolution, which introduced an implicit memory cache based on a dictionary.